rename to visualization part to visualization
